Allow swords to be picked up when camped
Decided to have the swords "expand" their range over time, up to a max
range. Now players can pick up swords without having to be as close to
them as before.

We need to telegraph picking up more effectively than we currently do in
order for this to be intuitive, but it should be fine overall.